feat: move test func to h
